Split up checkout ui to multiple files
